pyproject.toml
[project]
name = "sheetagent"
version = "0.1.0"
description = "AI-powered tool for analyzing Accounts Receivable (A/R) open posts lists and generating comprehensive aging reports."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "colorlog>=6.10.1",
    "fastapi>=0.128.0",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.8.0",
    "ipython>=9.9.0",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "recognizers-text>=1.0.2a2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "ruff>=0.9.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
